Remove forked rust dependency (#19)
This PR removes the forked rust dependency by implementing the following
two functions directly in smir_pretty (instead of in a forked rust
compiler):

-   visited_tys
-   visited_alloc_ids

This is done by extending the functionality of `LinkNameCollector` to
also collect `Ty` and `AllocId` mappings and renaming the combined
collector to `InternedValueCollector`.

EDIT: When running the panic example on rustc master, there is some
weirdness where a static item called:

`alloc::alloc::__rust_no_alloc_shim_is_unstable`

is generated which has no body. This static item appears to be
referenced from several different allocations (can be seen by chasing
pointers in allocations provenance table). If this funny static item is
findable on other machines, we should document this issue. It may well
be that this is an empty type that is never used at runtime, but I
haven't gone in to check this yet.

---------

Co-authored-by: Jost Berthold <jost.berthold@gmail.com>
Co-authored-by: dkcumming <daniel.cumming@runtimeverification.com>
Co-authored-by: Daniel Cumming <124537596+dkcumming@users.noreply.github.com>