Working with cargo (#47)
This didn't end up looking anything like what I had imagined. But I was
having a lot of trouble getting this working at all and I may have had
my environment set up in a way that was self defeating as I feel I
suddenly started having things work when trying the same thing. Not
sure.

But right now this is creating a `.stable_mir_json` directory at a given
directory or `~/.stable_mir_json` directory (linux only right now) and
that is populated with the binary for `stable_mir_json`, a file
containing the `LD_LIBRARY_PATH` that was used to create the binary, and
scripts to set the library path so that RUSTC can be overwritten when
calling cargo.

To test this out:
1. Fetch and switch to this branch
2. run `cargo build` and `cargo build --release`
3. run `cargo run --bin cargo_stable_mir_json -- $PWD
[optional-dir-path]` from the root of the repo (this creates
`.stable_mir_json`)
4. create or navigate to another cargo project you want to test
5. run `RUSTC=~/.stable_mir_json/debug.sh cargo build` (or the file that
`.stable_mir_json` was built in). Or for release
`RUSTC=~/.stable_mir_json/release.sh cargo build`

Fixes #39

---------

Co-authored-by: Jost Berthold <jost.berthold@gmail.com>