Extract discriminant information (#62)
As it turns out, the `SwitchInt` terminator uses `discriminants`
(computed using `RValue::Discriminant` ) but when an enum value is
created we only have the `VariantIdx` (`0..n` in source order, including
uninhabited variants).

This PR extracts a mapping of `VariantIdx` to discriminants (bytes
stored as `u128`) as type metadata for each `enum` that can be used in
the semantics to implement the `RValue::Discriminant`.