Add adt_def to struct and enum metadata (#64)
These `adt_def` of `struct` and `enum` metadata are required for a
reverse mapping `adt_def -> Ty` in the semantics.
The `AggregateKind` only carries the `adt_def` but we have to look up
the discriminant information (and potentially other metadata later) from
a `Ty` key.

Note that the `adt_def` is unstable, it changes with every compiler run,
so we delete the field in the golden tests.