Handling `FnPtr` signature internally to avoid `Binder`s (#71)
I think this is the correct way to access the function signature args
for a function pointer, even though it goes through internal. I checked
kani kani-compiler/src/codegen_cprover_gotoc/codegen/statement.rs and
this is what they are doing.

There was one failure `rust/tests/ui/issues/issue-26641.rs (exit 101)`
that uses a `RigidTy::Dynamic(Binder...)` but we haven't really looked
at details for `dyn` and `vtable` yet so this is moved to failing for
now.

Also improved the UI testing output but adding verbosity and some
metrics