Graph improvements (#72)
Some readability issues are fixed in this PR:

- `Deref` projections are now wrapped in parenthesis e.g.`(*4)`
- Constants are not longer separated by "`::`", e.g. `const :: u32` is
now `const ?_u32`, and `const :: std::ops::ExitCode` is now `const
std::ops::ExitCode`. I do not like that there is "`?`" but I couldn't
see a fast way to deal with the bytes in the `Allocation`, does anyone
know how to parse that into the integer?
- Changed `Rvalue::AddressOf` e.g. `&Mut (*3)` is now `&raw mut (*3)`,
which is less likely to be confused with a mutable borrow `&mut (*3)`
- Added a node with all the locals and type information. Some formatting
and placement issues with this. I would appreciate some feedback!

---------

Co-authored-by: Jost Berthold <jost.berthold@gmail.com>