Added `Span` information to JSON (#74)
Each item that comes out of the compiler has a `Span`, this PR now
includes those spans in the JSON output. Each `Span` in the JSON is an
`id: usize` that maps to a tuple `(path:String, col_i:usize,
row_i:usize, col_e:usize, row_e:usize)` to locate the source position.
The spans are sorted on the `id`, however I noticed some numbers are
missing - perhaps we need to investigate what we are missing.

---------

Co-authored-by: Jost Berthold <jost.berthold@gmail.com>