Add nix derivation (#96)
This adds a nix derivation for stable-mir-json.

In addition, the unit and integration tests can be executed via nix by
running `nix flake check`.

Also a nix formatter was added. It can be installed via `nix profile
install nixpkgs#nixfmt` and run via `make format` or `nixfmt **/*.nix`.

A test workflow was added that builds and tests the nix derivation for
both x86_64-linux and arm64-MacOS.

Finally, the derivation and its dependencies are published on the usual
k-framework cachix nix binary caches.

---------

Co-authored-by: JianHong Zhao <zhaojianhong96@gmail.com>