`TyKind` in `GlobalAlloc`s (#84)
`GlobalAlloc`s need to be decoded to their types, this can be done
easier by preprocessing the python (maybe in the future in the repo).
But for the python to preprocess it will go easier with the `Ty`
available in the JSON.

- Changed `AllocInfo` to be a struct that contains the `GlobalAlloc`
information, originally it was a enum that was a duplicate of
`stable_mir::mir::alloc::GlobalAlloc`. The struct helps with the JSON as
the field names come through.
- Changed the functions that get the `TyKind` from the allocs and provs
to work with `Ty`
- `collect_alloc` also takes a `Ty` as a parameter instead of a `TyKind`
now
- `AllocMap` is now collects the `Ty` also